# contracts/openapi/bootstrap.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IT Operations Hub Bootstrap",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "revokeSession",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Session revoked" },
          "default": { "description": "Canonical error" }
        }
      }
    },
    "/api/v1/health/live": {
      "get": {
        "operationId": "liveness",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Canonical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "operationId": "readiness",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Canonical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Canonical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/authorization/evaluate": {
      "post": {
        "operationId": "evaluateAuthorization",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "resource_type", "resource_id"],
                "properties": {
                  "action": { "type": "string" },
                  "resource_type": { "type": "string" },
                  "resource_id": { "type": "string" },
                  "scope": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Authorization decision" },
          "default": { "description": "Canonical error" }
        }
      }
    },
    "/api/v1/authorization/evaluate-batch": {
      "post": {
        "operationId": "evaluateAuthorizationBatch",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "501": { "description": "Batch contract pending" },
          "default": { "description": "Canonical error" }
        }
      }
    },
    "/api/v1/operations/{id}": {
      "get": {
        "operationId": "getOperation",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Canonical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "meta"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "details", "retryable"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "retryable": {
                "type": "boolean"
              }
            }
          },
          "meta": {
            "type": "object",
            "required": ["request_id", "correlation_id"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "correlation_id": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}